feat: safe ValidationError message access via .get_squeezed_message()
ValidationError.data and .message is provided by third party libraries,
and they can give a message of any length. E.g. jsonschema inserts its
input in all of its messages, which could be arbitrary big.

To be able to show these errors to the user, some pre-processing is
needed. The new method allows for squeezing these messages in a way,
that is least disruptive, and has special knowledge how to shorten
jsonschema messages.

It is definitely still a workaround, and ideally the libraries should
not yield unlimited messages.

Signed-off-by: Krisztian Fekete <1246751+e3krisztian@users.noreply.github.com>