fix: implement __lt__ for models still missing it
Since cyclonedx-python-lib has the tendency to use SortedSets to store
deserialized instances, this can lead to errors like these:

TypeError: '<' not supported between instances of 'ComponentEvidence'
and 'ComponentEvidence'

When loading a CycloneDX JSON SBOM using Bom.from_json().

This is due to the lack of 'less-than' operator, which is now
implemented with __lt__ for all models that were still missing it.

We implemented it this way since that's how it's done for all the
existing models, but an easier implementation would propably to add
functools' @total_ordering decorator to those classes.

Signed-off-by: Quentin Kaiser <quentin.kaiser@onekey.com>